+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/compare_aggregate_new.sql
@@ -1,0 +1,3 @@
+-- no diff
+CREATE AGGREGATE public.t1(*) (SFUNC = int8inc, STYPE = int8);
+CREATE AGGREGATE public.t2(bigint) (SFUNC = int8_avg_accum, STYPE = internal);